gh-payment/src/main/java/com/infnet/ghpayment/service/PaymentService.java
package com.infnet.ghpayment.service;

import com.infnet.ghpayment.dto.PaymentDTO;
import org.springframework.stereotype.Service;

@Service
public class PaymentService {

    public PaymentDTO processPayment(PaymentDTO paymentDto) {
        // Implementar lógica para processar o pagamento
        // Converter PaymentDto para entidade de domínio, salvar no banco de dados e retornar o PaymentDto resultante
        return null;
    }

    public PaymentDTO getPaymentById(Long paymentId) {
        // Implementar lógica para buscar um pagamento pelo ID e retornar o PaymentDto correspondente
        return null;
    }

    public void cancelPayment(Long paymentId) {
        // Implementar lógica para cancelar um pagamento pelo ID
    }
}
